developing-message-handling-structure
